+++ docs/whats_new.rst
@@ -54,6 +54,7 @@ Enhancements
 - Creating the `FilterBanklayer` module for new models (:gh:`656` by `Bruno Aristimunha`_ )
 - Including PytorchAudio as dependency and remove copied code (:gh:`661` by `Bruno Aristimunha`_)
 - Adding :class:`braindecode.models.FBCNet` (:gh:`657` by `Bruno Aristimunha`_ )
+- Adding :class:`braindecode.models.FBLightConvNet` (:gh:`657` by `Bruno Aristimunha`_ )
 
 Bugs
 ~~~~